Cleanup setup.py
