--- components/PyTorch/pytorch-kfp-components/dependencies.py
@@ -1,37 +1,0 @@
-#!/usr/bin/env/python3
-#
-# Copyright (c) Facebook, Inc. and its affiliates.
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-"""Setup configuration of  PyTorch Kubleflow Pipelines client side libraries."""
-
-def make_required_install_packages():
-  return [
-      "kfp>=1.6.1",
-      "torch>=1.7.1",
-      "torchserve>=0.3.0",
-      "torch-model-archiver",
-      "pytorch-lightning==1.3.2",
-  ]
-
-
-def make_required_test_packages():
-  return make_required_install_packages() + [
-      "mock>=4.0.0",
-      "flake8>=3.0.0",
-      "pylint",
-      "pytest>=6.0.0",
-  ]
-
-
-def make_dependency_links():
-  return []
