+++ components/PyTorch/pytorch-kfp-components/setup.cfg
@@ -1,0 +1,2 @@
+[metadata]
+description-file = README.md
