+++ components/PyTorch/pytorch-kfp-components/setup.py
@@ -15,63 +15,84 @@
 
 """Setup script."""
 
+import importlib
 import os
-import re
-import dependencies 
+import types
 
 from setuptools import setup, find_packages
 
-with open("README.md", "r", encoding="utf-8") as fh:
-    long_description = fh.read()
+def make_required_install_packages():
+  return [
+      "kfp>=1.6.1",
+      "torch>=1.7.1",
+      "torchserve>=0.3.0",
+      "torch-model-archiver",
+      "pytorch-lightning==1.3.2",
+  ]
 
-relative_directory = os.path.relpath(os.path.dirname(os.path.abspath(__file__)))
-# get version string from module
-with open(
-        os.path.join(os.path.dirname(__file__),
-                     "pytorch_kfp_components/__init__.py"),
-        "r",
-) as f:
-    version = re.search(r"__version__ = ['\"]([^'\"]*)['\"]", f.read(),
-                        re.M).group(1)
+def make_required_test_packages():
+  return make_required_install_packages() + [
+      "mock>=4.0.0",
+      "flake8>=3.0.0",
+      "pylint",
+      "pytest>=6.0.0",
+  ]
 
-setup(
-    name="pytorch-kfp-components",
-    version=version,
-    description="PyTorch Kubeflow Pipeline",
-    url="https://github.com/kubeflow/pipelines/tree/master/components",
-    author="The PyTorch Kubeflow Pipeline Components authors",
-    author_email="pytorch-kfp-components@fb.com",
-    license="Apache License 2.0",
-    long_description=long_description,
-    long_description_content_type="text/markdown",
-    extras_require={"tests": dependencies.make_required_test_packages()},
-    include_package_data=True,
-    python_requires=">=3.6",
-    install_requires=dependencies.make_required_install_packages(),
-    dependency_links=dependencies.make_dependency_links(),
-    keywords=[
-        "Kubeflow",
-        "ML workflow",
-        "PyTorch",
-    ],
-    classifiers=[
-        "Development Status :: 3 - Alpha",
-        "Operating System :: Unix",
-        "Operating System :: MacOS",
-        "Intended Audience :: Developers",
-        "Intended Audience :: Education",
-        "Intended Audience :: Science/Research",
-        "License :: OSI Approved :: BSD License",
-        "Programming Language :: Python :: 3 :: Only",
-        "Topic :: Scientific/Engineering",
-        "Topic :: Scientific/Engineering :: Artificial Intelligence",
-        "Topic :: Software Development",
-        "Topic :: Software Development :: Libraries",
-        "Topic :: Software Development :: Libraries :: Python Modules",
-    ],
-    package_dir={
-        "pytorch_kfp_components":
-            os.path.join(relative_directory, "pytorch_kfp_components")
-    },
-    packages=find_packages(),
-)
+def make_dependency_links():
+  return []
+
+def detect_version(relative_directory):
+    loader = importlib.machinery.SourceFileLoader(
+        fullname="version",
+        path=os.path.join(relative_directory, "pytorch_kfp_components/__init__.py"),
+    )
+    version = types.ModuleType(loader.name)
+    loader.exec_module(version)
+    return version
+
+
+
+if __name__ == '__main__':
+    relative_directory = os.path.relpath(os.path.dirname(os.path.abspath(__file__)))
+    
+    version = detect_version(relative_directory)
+
+    setup(
+        name="pytorch-kfp-components",
+        version=version,
+        description="PyTorch Kubeflow Pipeline",
+        url="https://github.com/kubeflow/pipelines/tree/master/components",
+        author="The PyTorch Kubeflow Pipeline Components authors",
+        author_email="pytorch-kfp-components@fb.com",
+        license="Apache License 2.0",
+        extras_require={"tests": make_required_test_packages()},
+        include_package_data=True,
+        python_requires=">=3.6",
+        install_requires=make_required_install_packages(),
+        dependency_links=make_dependency_links(),
+        keywords=[
+            "Kubeflow",
+            "ML workflow",
+            "PyTorch",
+        ],
+        classifiers=[
+            "Development Status :: 3 - Alpha",
+            "Operating System :: Unix",
+            "Operating System :: MacOS",
+            "Intended Audience :: Developers",
+            "Intended Audience :: Education",
+            "Intended Audience :: Science/Research",
+            "License :: OSI Approved :: Apache Software License",
+            "Programming Language :: Python :: 3 :: Only",
+            "Topic :: Scientific/Engineering",
+            "Topic :: Scientific/Engineering :: Artificial Intelligence",
+            "Topic :: Software Development",
+            "Topic :: Software Development :: Libraries",
+            "Topic :: Software Development :: Libraries :: Python Modules",
+        ],
+        package_dir={
+            "pytorch_kfp_components":
+                os.path.join(relative_directory, "pytorch_kfp_components")
+        },
+        packages=find_packages(where=relative_directory),
+    )
